+++ apps/dotcom/client/src/tla/components/dialogs/TlaCookieConsent.tsx
@@ -1,38 +1,28 @@
-import { useAuth } from '@clerk/clerk-react'
 import classNames from 'classnames'
 import { memo, useCallback } from 'react'
-import { useDialogs, useValue } from 'tldraw'
-import { handleConsentChange } from '../../../utils/analytics'
+import { useDialogs } from 'tldraw'
 import { useAnalyticsConsent } from '../../hooks/useAnalyticsConsent'
-import { useMaybeApp } from '../../hooks/useAppState'
 import { F } from '../../utils/i18n'
 import styles from './dialogs.module.css'
 import { TlaManageCookiesDialog } from './TlaManageCookiesDialog'
 
 export const TlaCookieConsent = memo(function TlaSidebarCookieConsent() {
-	const app = useMaybeApp()
 	const { addDialog } = useDialogs()
-	const auth = useAuth()
-
-	const user = useValue('user id', () => app?.getUser(), [app])
-	const isSignedIn = !!auth.isSignedIn
-	const consent = useAnalyticsConsent(isSignedIn, user, app)
+	const [, updateConsent] = useAnalyticsConsent()
 
 	const handleAccept = useCallback(() => {
-		handleConsentChange(true, isSignedIn, user, app)
-	}, [app, user, isSignedIn])
+		updateConsent(true)
+	}, [updateConsent])
 
 	const handleReject = useCallback(() => {
-		handleConsentChange(false, isSignedIn, user, app)
-	}, [app, user, isSignedIn])
+		updateConsent(false)
+	}, [updateConsent])
 
 	const handleCustomize = useCallback(() => {
 		addDialog({
 			component: () => <TlaManageCookiesDialog />,
 		})
 	}, [addDialog])
-
-	if (consent !== null) return null
 
 	return (
 		<div className={styles.sidebarCookieConsent} data-testid="tla-sidebar-cookie-consent">
