+++ apps/dotcom/client/src/utils/analytics.tsx
@@ -1,11 +1,9 @@
-import { TlaUser } from '@tldraw/dotcom-shared'
 import posthog, { PostHogConfig, Properties } from 'posthog-js'
 import 'posthog-js/dist/web-vitals'
 import { useEffect } from 'react'
 import ReactGA from 'react-ga4'
 import { useLocation } from 'react-router-dom'
 import { getFromLocalStorage, setInLocalStorage, useValue, warnOnce } from 'tldraw'
-import { TldrawApp } from '../tla/app/TldrawApp'
 import { useApp } from '../tla/hooks/useAppState'
 
 // Local storage key for analytics consent
@@ -98,26 +96,6 @@ export function configureAnalytics(
 			optedIn: consent,
 			user,
 		})
-	}
-}
-
-// Utility function to handle consent changes for both signed-in and signed-out users
-export function handleConsentChange(
-	newConsent: boolean,
-	isSignedIn: boolean | undefined,
-	user?: TlaUser,
-	app?: TldrawApp | null
-) {
-	if (isSignedIn && user && app) {
-		app.updateUser({ id: user.id, allowAnalyticsCookie: newConsent })
-		// Also update localStorage to keep them in sync
-		setStoredAnalyticsConsent(newConsent)
-		// Immediately configure analytics
-		configureAnalytics(newConsent, { id: user.id, name: user.name, email: user.email })
-	} else {
-		setStoredAnalyticsConsent(newConsent)
-		// Immediately configure analytics for signed-out users
-		configureAnalytics(newConsent, undefined)
 	}
 }
